Simplify display handling
Screen update is now done through sending messages to the terminal.
As a consequence, paint operations are not done at a fixed framerate,
but only when the screen pixels actually change.
The pixel buffer needs to be cloned to be sent to the terminal, but
it means we don't need an `Arc<RwLock<...>>` anymore.