Add support for stopping the CHIP-8 system
This is useful in scenarios where the hosting process can load new
images and we want to make sure no resources linger after the previous
image is stopped.

This is not obvious, since the synchronisation features assumed the
system was running for ever.

In particular, we need to stop the threads for the timers and keyboard
devices. This is done through a Drop implementation.

In addition, we need a way to interrupt the wait_for_key_press
blocking function.
To achieve this, we got rid of the condition variable, using a channel
instead, with the select macro that allows us to wait for several
channels at once. We also got rid of the dependency to parking_lot,
and switched to std mutexes instead.